git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@5315 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/USER-CG-CMM/Install.sh
@@ -2,8 +2,10 @@
 
 if (test $1 = 1) then
 
-  cp angle_cg_cmm.h ..
-  cp angle_cg_cmm.cpp ..
+  if (test -e ../angle_harmonic.cpp) then
+    cp angle_cg_cmm.h ..
+    cp angle_cg_cmm.cpp ..
+  fi
 
   cp cg_cmm_parms.h ..
   cp cg_cmm_parms.cpp ..
@@ -14,13 +16,16 @@ if (test $1 = 1) then
   cp pair_cg_cmm.h ..
   cp pair_cg_cmm_coul_cut.cpp ..
   cp pair_cg_cmm_coul_cut.h ..
-  cp pair_cg_cmm_coul_long.cpp ..
-  cp pair_cg_cmm_coul_long.h ..
+
+  if (test -e ../pppm.cpp) then
+    cp pair_cg_cmm_coul_long.cpp ..
+    cp pair_cg_cmm_coul_long.h ..
+  fi
 
 elif (test $1 = 0) then
 
-  rm ../angle_cg_cmm.h
-  rm ../angle_cg_cmm.cpp
+  rm -f ../angle_cg_cmm.h
+  rm -f ../angle_cg_cmm.cpp
 
   rm ../cg_cmm_parms.h
   rm ../cg_cmm_parms.cpp
@@ -31,7 +36,8 @@ elif (test $1 = 0) then
   rm ../pair_cg_cmm.h
   rm ../pair_cg_cmm_coul_cut.cpp
   rm ../pair_cg_cmm_coul_cut.h
-  rm ../pair_cg_cmm_coul_long.cpp
-  rm ../pair_cg_cmm_coul_long.h
+
+  rm -f ../pair_cg_cmm_coul_long.cpp
+  rm -f ../pair_cg_cmm_coul_long.h
 
 fi
